requirements.txt
# airhelper
PyYAML
CoAPthon3
pycryptodomex

# airserver
websockets
websocket-client

# dev
autopep8
